requirements.txt
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Flask==3.0.3
gevent==24.2.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-documentai==2.29.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.1
grpcio-status==1.65.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
MarkupSafe==2.1.5
numpy==2.0.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
setuptools==72.1.0
six==1.16.0
textdistance==4.6.3
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
zope.event==5.0
zope.interface==6.4.post2
unittest2
coverage
Flask-Testing